requirements.txt
asgiref==3.4.1
attrs==22.2.0
blinker==1.5
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
ConfigArgParse==1.7
dataclasses==0.8
django-appconf==1.0.5
django-sortedm2m==3.1.1
et-xmlfile==1.1.0
Flask==2.0.3
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==22.10.2
geventhttpclient==2.0.11
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==4.8.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
locust==2.8.7.dev8
MarkupSafe==2.0.1
msgpack==1.0.5
openpyxl==3.1.2
packaging==21.3
Pillow==8.4.0
pluggy==1.0.0
psutil==5.9.6
# psycopg2==2.9.8
# psycopg2-binary==2.9.8
py==1.11.0
pycparser==2.21
PyExcelerate==0.10.0
pyparsing==3.1.1
pytest==7.0.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.27.1
roundrobin==0.0.4
six==1.16.0
sqlparse==0.4.4
tomli==1.2.3
typing_extensions==4.1.1
tzdata==2023.3
urllib3==1.26.18
Werkzeug==2.0.3
zipp==3.6.0
zope.event==4.6
zope.interface==5.5.2
